Changed the Rhythm game loading
The rhythm game loading is now based on delta, where the notes are recorded with respect to delta. As we use the physics process, delta is approximately 1/60s, and as such, should be the most consistent precise time we have access to. Basing it off this solves many previous problems, and lets me clean up the code greatly.