Short file restructure + remapped Awake Now + Added map
Created a map for the navigation section of the game.
Setup a new file structure which will be further developed afterwards.
Remapped awake now on the higher end pc to dial in timings and the delay. Will see if hardware has any visible difference around  map timing, and  if needed, will add adjustable timing.